Using Java for processing the message
Ok, it's a slight blend of bash and Java, but yeah,

The commit-mgs script will now pass the branch and message to a Java program, which will process and spit out the commit message we need.

If the branch format doesn't match a YouTrack ticket, we just get the message. If no work time was added to the commit message, we just get the message